changes on 11/12
moved all css to a single style sheet. Everything seems to be in place,
still need a little work with margins and padding for a few elements.
Biggest issue is the social media tags, for some reason they aren’t
showing up on all the pages, and the placement isn’t the same on the
pages where they do show up